+++ include/clib/radeon3d_protos.h
@@ -46,5 +46,6 @@ BOOL Radeon3DSubmitFence(struct Radeon3DDevice *device, ULONG *fenceOut);
 BOOL Radeon3DAcquireLease(struct Radeon3DDevice *device,
                           struct Radeon3DLease *lease);
 BOOL Radeon3DReleaseLease(struct Radeon3DDevice *device, ULONG lastFence);
+BOOL Radeon3DHeartbeatLease(struct Radeon3DDevice *device);
 
 #endif